Locking Down DeleteItem Permissions

+++ app/backend/src/resolvers/Mutation.js
@@ -47,9 +47,17 @@ const Mutations = {
   async deleteItem (parent, args, ctx, info) {
     const where = {id: args.id};
     // Find the item
-    const item = await ctx.db.query.item ({where}, `{id title}`);
+    const item = await ctx.db.query.item ({where}, `{id title user { id }}`);
     // check if they own the item, or have permission
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermission = ctx.request.user.permission.some (permission => [
+      'ADMIN',
+      'ITEMDELETE'.includes (permission),
+    ]);
+
+    if (!ownsItem && hasPermission) {
+      throw new Error ("You don't have permission to do that!");
+    }
 
     // delete it
     return ctx.db.mutation.deleteItem ({where}, info);
